* devel: (56 commits)
  update version to 2.0pre9
  deb: pre9
  locales.conf: sanitize country codes
  locales.conf: add more locales
  tweaks: document Test Drive 2 problem
  fatfs: add missing fdpp ifdef [#1583]
  signal: inject timer before keypresses [#1580]
  pic: prevent timer from interrupting things [fixes #1580]
  term: fix regression of 2a97d388e [#1583]
  vgaemu: alloc extra pages for dirty_map
  dpmi: promote words to unsigned before lshift
  dpmi: fix misplaced braces
  fdpp: require 1.6 for country support
  fdpp: add country setting [fixes #1579]
  locales.conf: add some country numbers [#1579]
  sdl: fix warning
  fatfs: fix c++ build
  simx86: fix linkage of do_seq_start() in c++ mode
  various: underscore _min _max macros
  mhpdbgc: build with g++
  ...

+++ src/include/cpu.h
@@ -204,8 +204,8 @@ union g_reg {
 
 #define _LWORD(reg)	LO_WORD(_##reg)
 #define _HWORD(reg)	HI_WORD(_##reg)
-#define _LWORD_(reg)	LO_WORD_(_##reg, const)
-#define _HWORD_(reg)	HI_WORD_(_##reg, const)
+#define _LWORD_(reg)	((unsigned)LO_WORD_(_##reg, const))
+#define _HWORD_(reg)	((unsigned)HI_WORD_(_##reg, const))
 
 /* this is used like: SEG_ADR((char *), es, bx) */
 #define SEG_ADR(type, seg, reg)  type(LINEAR2UNIX(SEGOFF2LINEAR(SREG(seg), LWORD(e##reg))))
